Cleanup
Fixed some duplication in the HTML.  Looks like a copy and paste error.

Remove 'report' from url paths but left as relative paths.  Relative
paths don't point to Plone but only to files uploaded as part of the
theme.  This needs to be fixed.

Ran theme.html through HTML formatter to aid readability.